refactor: apply FSD structure and add linting
