--- src/App.css
@@ -1,101 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-}
-body {
-    background: linear-gradient(to right, #66549e, #88c3d8);
-}
-
-.app {
-    width: 100vw;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background: rgba(255, 255, 255, 0.5);
-    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.3);
-    border-radius: 10px;
-}
-
-
-
-.board {
-    width: calc(64px * 8);
-    height: calc(64px * 8);
-    display: flex;
-    flex-wrap: wrap;
-    box-shadow: 0px 0px 10px 2px rgba(0,0,0,0.75);
-}
-
-.cell {
-    width: 64px;
-    height: 64px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.cell img {
-    width: 48px;
-    height: 48px;
-    position: relative;
-}
-
-.black {
-    background-color: #663c1b;
-}
-
-.white {
-    background-color: #e89658;
-}
-
-.selected {
-    background-color: #d7672c;
-    box-shadow: inset 0px 0px 10px 2px rgba(0,0,0,0.75);
-}
-
-.available {
-    height: 12px;
-    width: 12px;
-    border-radius: 50%;
-    background-color: #22c022;
-}
-
-.lost {
-    height: calc(50vh - 60px);
-    padding: 30px;
-    margin-left: 50px;
-    background-color: lightgray;
-    box-shadow: 0px 0px 10px 2px rgba(0,0,0,0.75);
-}
-.cell.selected {
-    background-color: #d7672c;
-    animation: selected-cell 0.5s ease-out;
-}
-
-@keyframes selected-cell {
-    from {
-        transform: scale(1);
-    }
-    to {
-        transform: scale(1.2);
-    }
-}
-
-.available {
-    height: 12px;
-    width: 12px;
-    border-radius: 50%;
-    background-color: #22c022;
-    animation: available-cell 0.5s ease-out;
-}
-
-@keyframes available-cell {
-    from {
-        transform: scale(1);
-    }
-    to {
-        transform: scale(1.2);
-    }
-}
-
