+++ src/entities/player/Player.ts
@@ -1,0 +1,9 @@
+import { Colors } from '../../shared/const/colors';
+
+export class Player {
+  color: Colors;
+
+  constructor(color: Colors) {
+    this.color = color;
+  }
+}
