+++ src/index.tsx
@@ -1,9 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './app/App';
 
-ReactDOM.render(
-    <App />,
-  document.getElementById('root')
-);
-
+ReactDOM.render(<App />, document.getElementById('root'));
