--- src/models/Colors.ts
@@ -1,4 +1,0 @@
-export enum Colors {
-    WHITE= "white",
-    BLACK= "black",
-}
