--- src/models/Player.ts
@@ -1,10 +1,0 @@
-import {Colors} from "./Colors";
-
-export class Player {
-    color: Colors;
-
-
-    constructor(color: Colors) {
-        this.color = color;
-    }
-}
