+++ src/shared/assets/black-bishop.png
