+++ src/shared/assets/black-king.png
