+++ src/shared/assets/black-pawn.png
