+++ src/shared/assets/black-queen.png
