+++ src/shared/assets/black-rook.png
