+++ src/shared/assets/white-bishop.png
