+++ src/shared/assets/white-king.png
