+++ src/shared/assets/white-knight.png
