+++ src/shared/assets/white-pawn.png
