+++ src/shared/assets/white-queen.png
