+++ src/shared/assets/white-rook.png
