+++ src/shared/const/colors.ts
@@ -1,0 +1,4 @@
+export enum Colors {
+  WHITE = 'white',
+  BLACK = 'black',
+}
